Remove profile slice and handle fulfilled profile topic in middleware
