+++ src/redux/local/localSlice.ts
@@ -4,9 +4,6 @@ import { TaskbarItemConfig } from '@/panels/Menu/types';
 import { Uri } from '@/types/types';
 import { menuItemsData } from '@/windowmanagement/data/MenuItems';
 
-import { getProfile } from '../profile/profileMiddleware';
-import { ProfileState } from '../profile/profileSlice';
-
 export interface LocalState {
   sceneTree: {
     expandedGroups: string[];
@@ -66,21 +63,6 @@ export const localSlice = createSlice({
       state.taskbarItems = action.payload;
       return state;
     }
-  },
-  extraReducers: (builder) => {
-    builder.addCase(
-      getProfile.fulfilled,
-      (state, action: PayloadAction<ProfileState>) => {
-        Object.entries(action.payload.uiPanelVisibility).forEach(([key, value]) => {
-          const item = state.taskbarItems.find((item) => item.id === key);
-          if (item) {
-            item.visible = value;
-          }
-          return state;
-        });
-        return state;
-      }
-    );
   }
 });
