+++ src/redux/store.ts
@@ -11,7 +11,6 @@ import { flightControllerReducer } from './flightcontroller/flightControllerSlic
 import { groupsReducer } from './groups/groupsSlice';
 import { localReducer } from './local/localSlice';
 import { missionsReducer } from './missions/missionsSlice';
-import { profileReducer } from './profile/profileSlice';
 import { propertiesReducer } from './propertytree/properties/propertiesSlice';
 import { propertyOwnersReducer } from './propertytree/propertyowner/propertyOwnerSlice';
 import { sessionRecordingReducer } from './sessionrecording/sessionRecordingSlice';
@@ -40,8 +39,7 @@ export const store = configureStore({
     skybrowser: skyBrowserReducer,
     time: timeReducer,
     userPanels: userPanelsReducer,
-    version: versionReducer,
-    profile: profileReducer
+    version: versionReducer
   },
   middleware: (getDefaultMiddleware) =>
     getDefaultMiddleware().concat([listenerMiddleware.middleware]),
